+++ editor/editor-common/src/utils/browser.ts
@@ -32,7 +32,7 @@ if (typeof navigator !== 'undefined') {
   result.gecko = !ie && /gecko\/\d/i.test(navigator.userAgent);
   result.chrome = !ie && /Chrome\//.test(navigator.userAgent);
   result.chrome_version = parseInt(
-    (navigator.userAgent.match(/Chrome\/(\d{2})/) || [])[1],
+    (navigator.userAgent.match(/Chrome\/(\d+)/) || [])[1],
     10,
   );
   result.android = /Android \d/.test(navigator.userAgent);