+++ media/media-card/src/root/cardView.tsx
@@ -305,8 +305,6 @@ export class CardViewBase extends React.Component<
       forceSyncDisplay,
     } = this.props;
 
-    const { didImageRender } = this.state;
-
     return (
       !!dataURI && (
         <ImageRenderer
@@ -320,7 +318,6 @@ export class CardViewBase extends React.Component<
           onImageError={this.onImageError}
           nativeLazyLoad={nativeLazyLoad}
           forceSyncDisplay={forceSyncDisplay}
-          isImageVisible={forceSyncDisplay || didImageRender}
         />
       )
     );